requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.1
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1.1
django-heroku==0.3.1
django-rest-framework==0.1.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
gunicorn==20.1.0
idna==3.4
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.1
oauthlib==3.2.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.5.0
python3-openid==3.2.0
pytz==2022.2.1
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
tzdata==2022.3
uritemplate==4.1.1
urllib3==1.26.12
whitenoise==6.2.0
